# Retrieve resolved package versions in parallel