@dschaefer2 2025-01-07
> As always, I worry about missing dependencies when making an algorithm run in parallel. Are we sure nothing will try and fetch the results before they are ready?

Ah, I see we await until they're all done before we continue. Cool.