@dschaefer2 2025-01-14
I have to revert this. The checkoutRepository calls save the workspace but also update the state and there's no locking around the workspace state data. Until we ensure only one thread is changing the workspace state at a time, we can't do this in parallel.